requirements.txt
boto3
pygame
python-dotenv